Responsive landing page
Here these three are the javascript, css, and html code files and one image which used to build a beautiful responsive landing page and its very interactive. while scrolling the page the navigation bar color is changed. In the navigation bar the color is changed whilw selecting its because of hover style.